done for deleting confirmation

+++ app/(pages)/menu/option-group/update.tsx
@@ -588,37 +588,48 @@ const OptionGroupUpdate: React.FC = () => {
         <CustomButton
           title="Xóa nhóm lựa chọn"
           handlePress={async () => {
-            try {
-              const response = await apiClient.delete(
-                `shop-owner/option-group/${optionGroupModel.id}`,
-                {
-                  data: {
-                    id: optionGroupModel.id,
-                    isConfirm: true,
-                  },
-                }
-              );
-              Alert.alert(
-                "Hoàn tất",
-                `Đã xóa nhóm lựa chọn "${optionGroupModel.title}"!`
-              );
-              router.replace("/menu");
-            } catch (error: any) {
-              console.log(
-                error,
-                `shop-owner/option-group/${optionGroupModel.id}`
-              );
-              if (error.response && error.response.status === 404) {
-                Alert.alert("Oops!", "Nhóm lựa chọn không còn tồn tại!");
-                router.replace("/menu");
-              } else {
-                Alert.alert(
-                  "Oops!",
-                  error?.response?.data?.error?.message ||
-                    "Hệ thống gặp lỗi, vui lòng thử lại sau!"
-                );
-              }
-            }
+            Alert.alert("Xác nhận xóa", `Bạn có chắc xóa nhóm này không?`, [
+              {
+                text: "Hủy",
+                style: "cancel",
+              },
+              {
+                text: "Đồng ý",
+                onPress: async () => {
+                  try {
+                    const response = await apiClient.delete(
+                      `shop-owner/option-group/${optionGroupModel.id}`,
+                      {
+                        data: {
+                          id: optionGroupModel.id,
+                          isConfirm: true,
+                        },
+                      }
+                    );
+                    Alert.alert(
+                      "Hoàn tất",
+                      `Đã xóa nhóm lựa chọn "${optionGroupModel.title}"!`
+                    );
+                    router.replace("/menu");
+                  } catch (error: any) {
+                    console.log(
+                      error,
+                      `shop-owner/option-group/${optionGroupModel.id}`
+                    );
+                    if (error.response && error.response.status === 404) {
+                      Alert.alert("Oops!", "Nhóm lựa chọn không còn tồn tại!");
+                      router.replace("/menu");
+                    } else {
+                      Alert.alert(
+                        "Oops!",
+                        error?.response?.data?.error?.message ||
+                          "Hệ thống gặp lỗi, vui lòng thử lại sau!"
+                      );
+                    }
+                  }
+                },
+              },
+            ]);
           }}
           containerStyleClasses={`mt-2 w-full h-[48px] px-4 bg-transparent border-2 border-gray-200 bg-primary-100 font-psemibold z-10 border-secondary bg-white relative `}
           textStyleClasses={`text-[16px] text-gray-900 ml-1 text-secondary`}
